+++ packages/camera/camera_avfoundation/ios/camera_avfoundation/Sources/camera_avfoundation/Camera.swift
@@ -49,6 +49,7 @@ protocol Camera: FlutterTexture, AVCaptureVideoDataOutputSampleBufferDelegate,
   ///
   /// @param messenger Nullable messenger for capturing each frame.
   func startVideoRecording(
+    videoOutputPath: String?,
     completion: @escaping (Result<Void, any Error>) -> Void,
     messengerForStreaming: FlutterBinaryMessenger?
   )
@@ -129,4 +130,17 @@ protocol Camera: FlutterTexture, AVCaptureVideoDataOutputSampleBufferDelegate,
   )
 
   func close()
+}
+
+extension Camera {
+  func startVideoRecording(
+    completion: @escaping (Result<Void, any Error>) -> Void,
+    messengerForStreaming messenger: FlutterBinaryMessenger?
+  ) {
+    self.startVideoRecording(
+      videoOutputPath: nil,
+      completion: completion,
+      messengerForStreaming: messenger
+    )
+  }
 }