+++ packages/camera/camera_platform_interface/lib/src/platform_interface/camera_platform.dart
@@ -143,6 +143,7 @@ abstract class CameraPlatform extends PlatformInterface {
     int cameraId, {
     @Deprecated('This parameter is unused, and will be ignored on all platforms')
     Duration? maxVideoDuration,
+    String? videoOutputPath,
   }) {
     throw UnimplementedError('startVideoRecording() is not implemented.');
   }
@@ -152,7 +153,7 @@ abstract class CameraPlatform extends PlatformInterface {
   /// Please see [VideoCaptureOptions] for documentation on the
   /// configuration options.
   Future<void> startVideoCapturing(VideoCaptureOptions options) {
-    return startVideoRecording(options.cameraId);
+    return startVideoRecording(options.cameraId, videoOutputPath: options.videoOutputPath);
   }
 
   /// Stops the video recording and returns the file where it was saved.