+++ packages/camera/camera/CHANGELOG.md
@@ -1,5 +1,6 @@
-## NEXT
+## 0.13.0
 
+* Adds `videoOutputPath` support to `startVideoRecording`.
 * Updates minimum supported SDK version to Flutter 3.38/Dart 3.10.
 
 ## 0.12.0+1