+++ packages/camera/camera_platform_interface/CHANGELOG.md
@@ -1,5 +1,6 @@
-## NEXT
+## 2.14.0
 
+* Adds `videoOutputPath` support to `MethodChannelCamera.startVideoRecording`.
 * Updates minimum supported SDK version to Flutter 3.38/Dart 3.10.
 
 ## 2.13.0